rename displayProcessor file
