+++ packages/grafana-ui/src/utils/index.ts
@@ -1,7 +1,7 @@
 export * from './valueFormats/valueFormats';
 export * from './colors';
 export * from './namedColorsPalette';
-export * from './displayValue';
+export * from './displayProcessor';
 export * from './fieldDisplay';
 export * from './validate';
 export { getFlotPairs } from './flotPairs';
